Move Command to __init__.py
