+++ sceptre/providers/__init__.py
@@ -1,3 +1,5 @@
+import re
+
 from abc import ABC, abstractmethod
 
 from sceptre.exceptions import DuplicateProviderRegistrationError
@@ -98,3 +100,27 @@ class Provider(SceptreProvider):
             )
         else:
             self.__connection_manager = connection_manager
+
+
+class Command:
+    @abstractmethod
+    def execute(self):
+        """
+        Implements the logic for the command
+        """
+
+    @classmethod
+    def __init_subclass__(cls, provider=None, **kwargs):
+        command_type = cls.__to_camel_case(cls.__name__).split('_')[0]
+        command_name = cls.__to_camel_case(cls.__name__)
+        if not isinstance(provider, SceptreProvider):
+            raise TypeError("The provider {} supplied to the Command {} \
+                            is not of Type sceptre.providers.Provider".format(provider,
+                                                                              command_name)
+                            )
+        provider.command_registry.update({command_type: {command_name: cls}})
+        super().__init_subclass__(**kwargs)
+
+    def __to_camel_case(name):
+        s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', name)
+        return re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
